Move `migration` package from `ftue` to `messages:impl` module
