+++ features/ftue/impl/src/main/kotlin/io/element/android/features/ftue/impl/FtueFlowNode.kt
@@ -33,7 +33,6 @@ import dagger.assisted.AssistedInject
 import io.element.android.anvilannotations.ContributesNode
 import io.element.android.features.analytics.api.AnalyticsEntryPoint
 import io.element.android.features.ftue.api.FtueEntryPoint
-import io.element.android.features.ftue.impl.migration.MigrationScreenNode
 import io.element.android.features.ftue.impl.notifications.NotificationsOptInNode
 import io.element.android.features.ftue.impl.state.DefaultFtueState
 import io.element.android.features.ftue.impl.state.FtueStep
@@ -75,9 +74,6 @@ class FtueFlowNode @AssistedInject constructor(
         data object Placeholder : NavTarget
 
         @Parcelize
-        data object MigrationScreen : NavTarget
-
-        @Parcelize
         data object WelcomeScreen : NavTarget
 
         @Parcelize
@@ -113,14 +109,6 @@ class FtueFlowNode @AssistedInject constructor(
         return when (navTarget) {
             NavTarget.Placeholder -> {
                 createNode<PlaceholderNode>(buildContext)
-            }
-            NavTarget.MigrationScreen -> {
-                val callback = object : MigrationScreenNode.Callback {
-                    override fun onMigrationFinished() {
-                        lifecycleScope.launch { moveToNextStep() }
-                    }
-                }
-                createNode<MigrationScreenNode>(buildContext, listOf(callback))
             }
             NavTarget.WelcomeScreen -> {
                 val callback = object : WelcomeNode.Callback {
@@ -158,9 +146,6 @@ class FtueFlowNode @AssistedInject constructor(
 
     private fun moveToNextStep() {
         when (ftueState.getNextStep()) {
-            FtueStep.MigrationScreen -> {
-                backstack.newRoot(NavTarget.MigrationScreen)
-            }
             FtueStep.WelcomeScreen -> {
                 backstack.newRoot(NavTarget.WelcomeScreen)
             }
